extraire la logique de la liste des dossiers et la réutiliser dans la page mes dossiers
